#714 Simplified LodeRunner
Was able to use the HashMap approach for several Lode Runner BinLabels, but could not figure out a way to do it with KL div or the generalized latent vector. Might leave this challenge for Maxx. However, in the meantime, it should be able to use the old approach for KL Div and latent vars without trouble.