Easy to turn off/on individual components of enhanced CPPN genotype
The PicbreederTask now has three lines that can easily be uncommented to turn off specific aspects of the enhanced CPPN genotype: the scale, the rotation, or the translation. If you turn all of them off, then it is as if you had simply commented out the genotype line specifying the EnhancedCPPNPictureGenotype.

@annazwicker This code is available in the dev branch.