#736 Added unit test that passes
Basically, I was getting an index out of bounds because the previous test case in Mario never terminated the GAN process, so I just kept loading in a Mario level and trying to do Mega Man stuff on it. I have updated it to launch distinct GANProcesses for each game type and terminate upon completion.