Can I just use Java Serialization?
#348
Apparently, simply implementing the Serializable interface and using some built-in Java methods/classes takes care of serialization just fine. However, the only problem is that the format it saves the data in is incomprehensible. Also, I think that this approach would make the serialization very sensitive to any changes to the class being serialized. However, it just might work!

Should test more first, and also make a helper class that bundles the code relevant to serializing and deserializing