#348 Trying to remove wox, but...
There are a lot of references to the wox serialization, and some of these uses are in parts of the code that probably shouldn't be around anymore anyway. It will take some time to remove them all, so I added a command line parameter that allows for specifying of the serialization type. This allows me to keep using wox so that things work while gradually moving over to pure Java serialization.