#348 Added test of saving/loading archetype
I was thinking that the problem with resuming an interrupted evolutionary run might be that the serialization did not work properly on the archetype array, but the test seems to work, so I'm still not sure what the problem is. It's still good to have a test like this though.