Proposed change to how images are made
Specifically, to how HSB values are restricted. There is a commented out alternative to the way it has been done before. It works a bit better, but I don't like deviating from the way Picbreeder is supposed to work, especially since the enhanced genotypes see to work well in the interactive setting ... but we should check that.

@annazwicker Keeping you up to date.