Removed jfrog and reflection
My work in the less_maven branch eventually found the problem. I removed not just the jfrog entry from the pom file, but also the org-reflections entry which was using jfrog in some way. Interestingly, I don't seem to have needed to sacrifice any functionality as a result of removing org.reflections, but we'll see if something ends up breaking down the line. As is, everything compiles and the unit tests pass. #672