#741 Remove links that are associated with problem neurons.
If any nodes are found along a path from an output neuron, those neurons are eliminated.  This at least helps fix the problems encountered when trying to view the images, but now some of the colors associated with these images are wrong.  Not sure if this has to do with removing these neurons or if there was a problem with the color all along.