#737 Testing out batch file. Code Crashed from easy-to-fix exception.
I added a batch file to test out the Level Path Bins, and things seemed to be going fine until I encountered an Unsupported Operation Exception. However, it was easy to fix ... I just had to supply the number of genome variables per segment, which was the number of latent variables plus the number of non-latent variables.