"Fixed" Mega Man level length change as well
This had the same issue as the recent commit for Mario. It wasn't crashing, but I think that is just because the default setting was the longest possible length, meaning that reducing the length would not lead to array index out of bounds, but it would mean that genomes had the wrong length. This is fixed now.