Picture Innovation Solved #753
The reason for the out of memory error was my attempt to optimize by parallelizing the initialization of the archive. Now there is a command line parameter that determines whether or not this is parallel, and if it is false then the code works ... although I also had to fix a synchronization issue at another part of the code that is still parallel.