#635 Remove Java Autoencoder
Although I will still be keeping DL4J around for the innovation engine, I want to get rid of DL4J stuff that's not working. The Java-based DL4J Autoencoder for the PictureTargetTask just doesn't work, so I'm deleting that code and closing this issue.

I also fixed a minor problem with the PictureTargetTask (error in the file path for a Python script)