#743 Fixed Mario CPPN evolution
I think this is fixed. This approach does not depend on segment-based fitness, do now when the MarioTask can't get a valid segment-based calculation, it will just ignore the resulting null values. Since those components won't contribute to the fitness calculation for the plain CPPN generated levels anyway, it is not a problem. I need to double check that nothing with GAN-based level generation is broken, but this seems to work.