#737 Only adding connected binStrings to hashMap
Only added to hashmap the labels that were connected. This was done by filling a boolean array with true and false (true to match with a 1 from the binary string, false otherwise). Also made recursive explore method that checks adjacent 2D array cells to see if they are connected. To finally determine if it is connected, I then checked for remaining "true" values in the array.