Plane distance inputs "work" but are calculated wrong
The work @alejmedinajr did almost completed issue #38 but not yet. I made some changes. I removed an unnecessary exception call in the 3D Util class. I also shortened all of the input name labels so the check boxes would fit on the interface.

The problem with the new inputs is that they are not properly centered. They are treating the corners of the shapes as being at (0,0,0) instead of the actual centers.