#752 Made images square, renamed x/y translation
The button width and height were different, which was causing the non-square images. This has been fixed ... images are all square now because Parameters.parameters.integerParameter("imageSize") is used for width and height everywhere.

Also, it turns out that x/y translation were NOT mislabeled. Rather, the direction of translation also depends on rotation, so things only seemed to be moving the wrong way because they were rotated somewhere close to 90 degrees or 270