Show solution paths for CPPN2GAN in MegaMan
Not exactly related to #608 but I noticed this issue while working on it. The level View button for CPPN2GAN was different from the one for plain GAN level breeder. It wasn't showing A* solution paths, but now it does.

There is probably an opportunity to reduce code repetition here, but I'm going to leave this as-is for now.