#770 Implemented flying machine fitness function
There was a method that was extracted from diversity block fitness that was used for both itself and the block movement fitness. In block movement fitness it is supposed to compare two arrays that represent the original structure position and the updated blocks position if they moved or not. Still needs to be tested