#770 Doesn't seem to be working correctly
I started ambitious and tried testing a different kind of flying machine but not getting the fitness score that I expected. Then I tried on a smaller scale and it works for the redstone and piston but not the redstone with two pistons. It keeps getting a fitness of 5 instead of 1. A lot more blocks are being added to the updated positions list and I'm not sure why.