#770 Test works for smaller examples
For examples with one to two pistons and one to two redstone blocks, it seems to be performing as it should. Now that I am testing with an actual flying machine based on the example on evocraft provides, it is behaving oddly. Just by comparing the original position to the updated position lists, I am noticing that some of the blocks in the updated list are mislabeled simply because it was changed by a piston. Many of the blocks like redstone and slime are not named as such but are called "PISTON_EXTENSION", which doesn't make a lot of sense. It also doesn't seem to be listing the air blocks (which seem to still be added to the fitness score) that appear when a piston goes back to not being extended. I am going to conduct smaller tests to see if my assumptions are right.