#770 Fitness function seems to work
Based on the tests that were made, it seems like the fitness function is able to properly calculate the fitness score whether the shape is moving or not. In testing I made the error be within 0.5 of the expected value, so I'm not sure if this could cause problems in the future.