#785 Print blocks in archive dir
Whenever a shape is created that it placed in the Minecraft world archive, the code now saves a text file listing the blocks in the shape. This is useful in general, but also helps in troubleshooting. However, the files are only appearing for shapes generated after initialization, and all of these files are correct. This indicates that the source of the problems comes from shapes generated during initialization.