#783 New genotype made
The new genotype for evolving the list of blocks has been created but still have exceptions being thrown. The discrete ceiling was initialized to be the size of the command line parameter made (currently has the label as amount of blocks to evolve but I think this should change to reflect what it actually does) and each value gets initialized with the value of the length of the block set that is being used.