#785 Fixed error in width calculation
Width and other range values were actually being calculated wrong. The problem occurred when the shape was empty. However, this was hard to catch because the list that was read would originally have AIR in it, which was filtered to an empty list, but this skipped a case that was checking for an actually empty list as opposed to one with AIR