#794 Added empty space offset to to MinecraftShapeTask
MinecraftShapeTask was behaving differently from the loner task version used for MAP Elites. This version was generating shapes at the minimal corner of the reserved empty space rather than in the center, which has a change to cause inconsistences with a fitness function like this one.

This could affect some of @MuellMark 's unit tests for #780 as well, but if anything the new util class and helper functions should just make everything easier.