#792 Works with longer wait time
The same set of block that make up the flying machine were spawned in again and the wait time before it checks for the updated center of mass was made longer to ensure that the flying machine would completely move out of the checked area. It passed so it seems like it works.