#775 Created new class
I created the new class in the mutation package called ConvertMinecraftCPPNtoBlockVectorMutation. I also put in some assert statements related to what we want to check prior to the code of this new class executing.

Also the method has some of the basic things as a starting point for what I will be doing.