#827 Doesn't work, but interesting
Piston heads are part of the problem. Piston heads can extend before the initial read and retract afterwards, which makes it seem like blocks have flown away.

However, it seems like there is no unique block type for extended sticky pistons, and they simply use the label STICKY_PISTON which makes it impossible to distinguish the extensions from the sticky pistons themselves ... at least, that is my hypotheses. Not sure yet, and if true, not sure what to do