#827 Half blocks need to move before flying
I couldn't figure out a way to properly ignore certain blocks. The current calculation filters out piston_heads from the block count (not the center of mass) which helped a bit, but I was still getting shapes that lost 3 or so blocks early on. Now, the number of blocks that need to "leave" before the shape counts as flying equals half the total number of blocks in the shape. This could be vulnerable to similar issues, but it certainly seems as though a shape is more likely to be flying once half of the blocks are gone.