#819 Taking too long
Evolving Zelda dungeons with the restricted scheme was taking WAY too long. In fact, I terminated it before finishing. I guess it keeps generating a lot of big levels with long A* paths, whereas the full version makes lots of crappy small levels. Is that true?

Need to investigate, but for now I've halved the number of individuals to generate to speed things up. Maybe I should add parallelism too?