Massive revision of how fitness functions work
Decided that all fitness functions should simply have access to the original block list from the generator rather than relying on the client to correctly read the world. The list from the generator will always be correct, and will precede any mechanical movement, so it should always be available as a reference point, especially for change in center of mass. This changes a lot of code, but most of it should be refactoring. The only fitness function that really changes is the change in center of mass, and I am still troubleshooting its unit tests, which are broken now.