#830 Inconsistency between machines
This test ran with a reliable score on my home machine, but on a lab machine the resulting score is reliably a different value. I guess there is variation from machine to machine based on hardware resources or software environment.

Hopefully, modifying the test to accept a tight range of values like this will be sufficient.