For proper experiments, no interactive with archive
We have recent evidence that machines in the archive can crash into neighbors, disrupt their associated emerald/diamond blocks, and cause changes in the archive as a result. Also, it is extra wasted computation. Just save the effort to evolve faster.

Hopefully the saving of block lists will be enough to check the final results.

Might also be nice to place the archive in the world at the end.