#835 Problem with orientation index
I actually did find a fairly annoying problem in the VectorToVolume encoding. It probably won't affect success too much and I'm not even sure it will directly affect this issue with the WHD scheme, but it is definitely an error, and it means I have to re-run ALL of the accumulated experiments so far. Argh

Going to make unit tests first though