#839 Improvements to change in center of mass
Hopefully this doesn't break anything. I think it is safe to make the requirement on the number of departed blocks more lenient, since the other fix of basing the initial count on what was generated rather than read was the biggest issue. Also became aware of a case where a shape could fly away, but leave behind oscillating blocks rather than still blocks.

I think this is better, but will require close observation.