#842 Farthest center of mass
Instead of looking at the last center of mass, the one farthest from the initial point is considered for tracking the precomputed results. This should improve the naming of text files (knowing the direction they flew in) and mast also fix the directional binning scheme. Also, it should not affect the fitness scores (which are based on accumulation), so this change will be compatible with previous results.