Do not restrict during init
What if every randomly initialized genotype is outside of the restricted range? There would be no starting point for evolution. This modification allows us to ignore the restriction during initialization. This assures that we will have a population to evolve with, but could create problems. Those individuals will always be in the archive, and we may never find anything in the restricted range when sampling from them. Still, we need to try something.