Restricted range before init still causing problems
Previous commits related to this did not work. The setting discardFromBinOutsideRestrictedRange needs to be set to false before evaluation, and only restored to true after evaluation and placement in the archive is complete. It was clear that we were not doing this correctly before, but I though that this modification would fix the problem. However, the code still crashes. So, this doesn't work yet ... a closer look is needed.