#874 Launch and terminate server just once
So, it seems that the problem is that launching and terminating a server with one set of unit tests, such as ChangeCenterOfMass, makes it so the next set of tests that try the same time won't be able to communicate with the server. There may be some way around this that involves waiting longer to properly assure the first server terminates before the second one starts, but a quicker solution was to add a way to prevent the server from terminating until all tests that need it have finished.

This led to the creation of the MinecraftServerTestTracker class, which is not a great thing to have to remember to update, but it seems to work.  @JoannaBlatt