#874 Synchronization and reduced clear area
I also had to make the clear area smaller to make several tests run. They would freeze due to the very large space. I sort of wonder if this was the actual cause of the problem now, but I don't want to switch everything back now that it works (maven install works!). I did add synchronization to the access of the test server count though, just in case.