#884 Where fitness is evaluated
Found spot in code that would not work right if multiple timed evaluation fitness functions were used. Will now throw an exception if an attempt is made. Then I made a batch file that causes that exception to be thrown. Once code is added to handle the special case of multiple times functions, the batch file will work