#884 TimedEvaluationMinecraftFitnessFunction can evaluate multiple functions now
Modified TimedEvaluationMinecraftFitnessFunction so that there is an option to accept multiple TimedEvaluationMinecraftFitnessFunction instances and get the scores for each one as the result of a single evaluation period.

This functionality is not taken full advantage of anywhere. The standard behavior is for a list of just the one function to be made, resulting in a result array with only one value in it. However, the method for evaluating multiple functions is public and can be called from elsewhere when needed.