Removed redundant block spawning
I'm pretty sure spawning blocks here is not needed, since the timed evaluation fitness functions spawn blocks as needed. This is a starting point for #894, but might also solve #867. I think that we were generating a shape, immediately clearing it, and then generating it again for evaluation ... but the initial spawn command would place the blocks in a different place if the shape was too low to the ground. Need to check to verify.