#900 Remember parent scores
In order to speed things up, I found a more general issue to address beyond Minecraft. The Mu+Lambda scheme should remember parent scores and reuse them! At least, it should be an option. This would typically not be done in noisy domains, and although Minecraft has some noise in evaluation, most of what we're looking at is consistent enough to make it worth remembering the scores.