#890 Test fails: thinks it is flying
This new test creates an oscillating machine that explodes. The fitness function gives it max fitness. It thinks it is a flying machine, but it is not. The reason for the problem is pointed out by a TODO note in ChangeCenterOfMassFitness. There is also another TODO note there pointing out a likely problem, though that is not confirmed yet with a unit test. @JoannaBlatt