Cleaned up ChangeCenterOfMassFitness
Made major changes to ChangeCenterOfMassFitness that fix most of the unit tests. Specifically, the tests should not end early if the shape is oscillating. Although such a result is short of a flying machine, it still deserves a complete evaluation. Only a completely still machine should cause early termination.

However, testTNTOscillating seems to still fail. Also, it is not clear if this actually solves the problem of exploded machines earning max fitness.

#890