#890 Fixed TNT test
Had to add checks to make sure the shape actually moved a sufficient amount, but this led to several other fixes. The code is cleaner now, though there is a bit of code that is almost duplicated, which might be nice to clean up later.

Also, testChangeInPosition seems to be broken now