#891 #919 Massive clean up of MOME logging
I cleaned up the MOME logging a lot, and also fixes several errors regarding which file was being plotted.

More logs are needed, and given the structure we have here now, I'm wondering if we still need a separate MOMELog class. Don't go forward on #919 yet. Simply create more logging as described in #891. It might also be time to pursue #917