Don't clear with glass
Something seems to be slowing down the evolution runs quite a bit. Clearing with glass seems like an obvious candidate, so I've turned it off by default. Merge my branch so that we won't be using it with future runs. However, keep an eye on things to look for problems.

@JoannaBlatt @TjRaffert