#908 Flying Fitness might work now
Added a lot of frustrating restrictions. I think the fitness was being overly strict before and disqualifying certain flying machines, but it was also too lenient and allowing a shape that exploded. That exploded shape is now disallowed, but the margin seems too narrow.

Being a flying machine now depends on having a center of mass that moves enough, and on having a farthest block that is far enough away. More testing is needed, and I may need to check/add unit tests.