#912 Start of approach to tracking component scores
I've tweaked the code to eventually make it possible to track the component scores of a weighted sum. Currently, this code does nothing useful. A few methods that returned a double array now return a Pair of two double arrays, but he second array is currently always of length 0.

The goal is to add more code so that this array can contain the component scores we need. The trick will be, in the calculateFitnessScores method of MinecraftShapeTask, to detect the weighted sum functions and call a new different method instead of fitnessScore. Instead of fitnessScore, we need something that returns the combined score and its components.