#910 working on synchronization
I have a strong feeling that the errors we are seeing are synchronization issues at this point, and I'm hoping this latest addition will fix the problem. There is a risk that it will slow down execution a bit, but hopefully not. I only lock the whole archive when a bin is being replaced/updated, but when a bin is just being checked, then only that one bin population gets locked.