#928 Currently over synchronized but functional
currently have created a near fully whole archive synchronization, but could be made more fine grained if necessary. The issue is that when a new bin is added it is empty and if another thread is doing logging it's functions could come across the empty bin and this causes issues. Currently need to see if it is significantly slower or not.