#932 Actually control how much extra space is cleared
Shockingly, there really was no clear way to make a shape clear additional space around it before each evaluation. The EMPY_SPACE_OFFSET was hardcoded. However, it is now controlled by a parameter. However, that parameter still isn't in the batch files since I just made the default value sufficiently big. Still, it is easy to change now if we want to.