#940 Don't rely on change in center of mass that much
Allows for a smaller change in center of mass to trigger one of the conditions for being considered a flying machine. The farthest block still needs to be very far away though, which should prevent false positives. Allowing a smaller center of mass change should benefit flying machines that leave more blocks behind, but it's possible we don't want this.