More space for missiles
I found that I could create more space in the evaluation area for missiles if I reduced the space in the y dimension. Since we only want missiles flying horizontally toward targets, this should work fine. However, we should not mess around with the space measurements for big flying machines, which we want to encourage to fly in all directions.
@TjRaffert