#943 Found place to scale, still need to do it
Turns out MAP Elites solved the lowest score problem with a command line parameter offsetting the fitness values. Using that won't make sense for all of the otherStats, but there is now a way I can pass in a parameter. I think it can just be the negation of the minScore, but want to wait until I have more time to implement and test