#946 created an assert to check shape corner
since it doesn't seem like shape corner is checked to make sure the block list aligns with it I have added an assert to check for this. It doesn't seem like it ever checks that the corner matches anywhere, and I'm a little unclear about why we even pass the corner instead of calculating it whenever a block list is passed.