#946 assert was breaking tests
This assert was breaking tons of unit tests, and it was based on a flawed assumption. Although it is true that some of the unit tests are not perfectly aligning the shape with the evaluation corner, this is not strictly necessary as long as they are close. However, in terms of bigger picture of evolving shapes, it is possible for a shape to not have any blocks within the minimum dimensions of the shape corner, which would make the shape corner not match the min corner.