Simple example in the complex group
This is a simple batch file that models an older GECCO one. Just evolved small flying machines with the direct encoding. Need to have some thorough tests that verify that this direct representation is really better. Probably need to add more batch files for small flying machines to recreate old results with the new encoding and settings.