Removed DL4J models
All DL4J models, which were mainly associated with Image Net innovation engines and HyperNEAT/Hybrid, were removed. Some code for backprop in the Tetris Task had to be removed, but it never really worked anyway. Code for generating DL4J networks from CPPNs never resulted in anything interesting.

The only thing that was sad to loose was the image net models (VGG, etc) that were part of innovation engines for both shapes and pictures. Somehow, error free code still exists in the innovation engine package, but I suspect a lot of it won't survive.

Also, the DL4J libraries are still in Maven, so some code depending on that is still present.