#952 fixed other stat issue
The Zelda CPPN2GAN experiments were using two objectives in the batch file. Only one was being used for selection (also, count vs percent is equivalent for selection), but the new multiobjective code interpreted this as an error for MAP Elites. Also, the otherStats for Zelda were being tracked incorrectly, in that 6 were expected but only 4 were logged. This error was caught by new error checking code. Also, the new code wants to track Pareto front and hypervolumes of these 6 stats, but that does not make sense here, so I added a parameter to disable the otherStat hypervolume logging.