MoleculeProcess does not work with FORTRAN code yet
Not completely sure why it does not work, but it seems that when executing the compiled Fortran code in PowerShell that it exits immediately without waiting for input. I'm only able to run the code from MSYS2 so far, so how do I change the code to make it run correctly in PowerShell?