Maybe fixed MUTATE5
I found an input sequence that causes the problem in the code. This sequence is in ProblemMutationSequence.txt. This can be used to recreate the error.
This made me realize that the code was not frozen at that point, but it did not produce any output, and thus the Java code was waiting to "receive" a line of output that would not come.

So, I looked for a case in MUTATE5 where there was no output and found one. I believe I've fixed the code in SMILESMutate.f that is an error case, and added an output of X.

The next chance I get, I need to recompile SMILESMutate.f in order to properly test.