Fixed Easy save of SMILES
Keeping the array list of mutation operators inside of the genotype, even as a static entity, was breaking Easy.save. So, I simply moved it out into some other class and I retrieve the information via a method call. This allows lengthy evolution.

However, there are still mutation operators that do not work. Fixing the change atom type mutation is top priority, though there is also a problem with fitness calculation on genotypes with rings.