Tracking errors
The Fortran program sometimes says the melting point is higher than the boiling point, which makes no sense. These are circumstances when the model used for calculating these quantities cannot be trusted, so this now results in BAD values of 999 which are interpreted to assign low fitness.

Also, there seems to be an issue with the archive structure that I'm trying to troubleshoot. Perhaps some molecules have more than 10 of a given atom type, which I admit I'm not handling yet.