Reduce output
Running MAP Elites creates lots of unnecessary logging information. This is an attempt to reduce that. Scores for individual elites were storing one line for every bin in the archive for backwards compatibility with the old Innovation Engine that allowed one elite to occupy several bins. That is uncommon now, so less information should be logged.