Troubleshooting
#961
It's ok to not add a ring if there is no room. No need to crash.

I guess it's ok to have two separate branches off of an atom as well, at least for Carbon.

I noticed that a ring can be added on an Oxygen which already has two bonds. That shouldn't happen. There is also a problem that double branching causes with the getAtomBondCount method